requirements.txt
python-telegram-bot==21.7
APScheduler==3.10.4
psycopg[binary,pool]==3.2.13
python-dotenv==1.0.1
